+++ src/test/java/racinggame/model/CarsTest.java
@@ -1,4 +1,4 @@
-package racinggame.model.Cars;
+package racinggame.model;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -6,7 +6,7 @@ import org.junit.jupiter.params.provider.NullAndEmptySource;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 import static org.assertj.core.api.AssertionsForInterfaceTypes.assertThat;
-import static racinggame.model.Cars.Cars.INPUT_ERROR_MSG;
+import static racinggame.model.Cars.INPUT_ERROR_MSG;
 
 class CarsTest {
   @Test
